Audio: use double for seeking
At moderately large positions (e.g. 6 minutes at 48 KHz), float32 is only
precise to 2 samples.

+++ src/audio/alstream.h
@@ -66,7 +66,7 @@ struct ALStream
 	AtomicFlag threadTermReq;
 
 	AtomicFlag needsRewind;
-	float startOffset;
+	double startOffset;
 
 	float pitch;
 
@@ -96,13 +96,13 @@ struct ALStream
 	void close();
 	void open(const std::string &filename);
 	void stop();
-	void play(float offset = 0);
+	void play(double offset = 0);
 	void pause();
 
 	void setVolume(float value);
 	void setPitch(float value);
 	State queryState();
-	float queryOffset();
+	double queryOffset();
 	bool queryNativePitch();
 	int getNumberOfComments();
 	char** getComments();
@@ -113,7 +113,7 @@ private:
 	void openSource(const std::string &filename);
 
 	void stopStream();
-	void startStream(float offset);
+	void startStream(double offset);
 	void pauseStream();
 	void resumeStream();
