+++ src/audio/audiostream.h
@@ -128,11 +128,11 @@ struct AudioStream
 	void play(const std::string &filename,
 	          int volume,
 	          int pitch,
-	          float offset = 0,
+	          double offset = 0,
 			  bool doFadein = true);
 	void stop();
 	void fadeOut(int duration);
-	void seek(float offset);
+	void seek(double offset);
 
 	/* Any access to this classes 'stream' member,
 	 * whether state query or modification, must be
@@ -143,7 +143,7 @@ struct AudioStream
 	void setVolume(VolumeType type, float value);
 	float getVolume(VolumeType type);
 
-	float playingOffset();
+	double playingOffset();
 	int getNumberOfComments();
 	char** getComments();
 	void setLoopPoints(int newLoopStart, int newLoopLength);
